Refactor Processing logic for testing
Moved old versions of logic to CruiseProcessing.ReferenceImplimentation project. Moved logic for processing into CruiseProcessor class. Reworked some of the processing logic to so that its only loading data for the stratum being processed. Before it would load data for all strata and then filter its as needed. This created potential for bugs, of which some were found. Also some renaming of variables and methods for clarity.